Stop SSH from sending the locale environments
Most servers set the default language as `en_US`, i.e. "English in
United States". If local language is set to `en_US` as well, there is no
need to send the configuration; otherwise, the local language (such as
Chinese) may not be available in remote machine. It's safer to configure
that in remote user profile such as `.bashrc`.

Note that in lines with whitespace only, the whitespaces should be
preserved.